Refactor React components
