+++ apps/electron/src/renderer/pill.html
@@ -11,6 +11,6 @@
 
   <body>
     <div id="root"></div>
-    <script type="module" src="/src/pill.tsx"></script>
+    <script type="module" src="/src/components/pill.tsx"></script>
   </body>
 </html>
