+++ apps/electron/src/renderer/src/onboarding.tsx
