+++ apps/electron/src/renderer/src/pages/dictionary.tsx
