+++ apps/electron/src/renderer/src/pages/vocabulary.tsx
